Refactored directory and module names
